Major Bug Fix:   Unscheduled bug fix for half-way bounce-back condition which was adding mass to the system. Other unscheduled changes are: -- Added Utility class and better log file writing and deleted the old generic operations file -- Temporarily fall back to copy exit conditions for stability -- MRT constants can now be edited from the definitions file rather than the source -- Streaming routine completely rewritten to make it easier to add streaming exclusions Sorry for not tracking this on a branch!

+++ LatBo/globalvars.h
@@ -11,7 +11,7 @@ extern const double w[nVels];				// Weights
 extern const double cs;						// Lattice sound speed for lattice
 #ifdef USE_MRT
 extern const int mMRT[nVels][nVels];		// MRT transformation matrix
-extern const float mInvMRT[nVels][nVels];	// Inverse of MRT transformation matrix
+extern const double mInvMRT[nVels][nVels];	// Inverse of MRT transformation matrix
 #endif
  
 #endif
